pyproject.toml
[tool.poetry]
name = "peq"
version = "0.1.0"
description = ""
authors = ["peresypkinmi <peresypkin.ma@sibvaleo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.3.1"
pytest-voluptuous = "^1.2.0"
allure-pytest = "^2.13.2"
curlify = "^2.2.1"
python-dotenv = "^1.0.0"
selene = "2.0.0rc2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

